requirements.txt
transformers
accelerate
huggingface-hub
langchain
langchain-community
langchain-core
python-dotenv
torch
sentence-transformers
faiss-cpu
youtube-transcript-api
streamlit
